+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Karafka Framework Changelog
 
 ## 2.5.1 (Unreleased)
+- [Enhancement] Support past `dispatch_at` times with `jitter: 0` in the OSS Karafka to support ActiveJob continuation.
+- [Enhancement] Use direct topic dispatches when `dispatch_at` is used for past times to bypass Scheduled Messages flow.
 - [Enhancement] Support immediate error raising with `auto.offset.reset` set to `error`.
 - [Enhancement] Don't create not needed dirs in the non-Rails setup template.
 - [Enhancement] Improve printing of TTIN to separate threads